+++ tests/bootstrap.php
@@ -1,0 +1,6 @@
+<?php
+
+require_once __DIR__ . '/../vendor/autoload.php';
+
+error_reporting(E_ALL ^ E_USER_DEPRECATED );
+